% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_colors.R
\name{scale_color_discrete_tpl}
\alias{scale_color_discrete_tpl}
\title{ggplots with TPL brand colors}
\usage{
scale_color_discrete_tpl(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_discrete()}
}
\value{
A ggplot layer that colors discrete scales with brand colors
}
\description{
Function that sets custom TPL brand color in discrete-valued plots. It works just like the built-in ggplot layer scale_color_discrete()
}
