% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_colors.R
\name{scale_fill_continuous_tpl}
\alias{scale_fill_continuous_tpl}
\title{ggplots with TPL brand colors}
\usage{
scale_fill_continuous_tpl(color, reverse = F, ...)
}
\arguments{
\item{color}{String that specifies which TPL brand color to use for gradient color scheme. Must be one of 'darkblue', 'teal', 'green', 'yellow', 'orange'}

\item{reverse}{Boolean that assigns the highest values to the lightest hue when set to TRUE (default=FALSE)}

\item{...}{Additional arguments passed to scale_fill_continuous()}
}
\value{
A ggplot layer that fills continuous scales with brand colors
}
\description{
Function that sets custom TPL brand color in continuously-valued plots. Other than customized parameters, it works just like the built-in ggplot layer scale_fill_continuous()
}
