% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tpl.R
\name{theme_tpl}
\alias{theme_tpl}
\title{Customize a ggplot theme}
\usage{
theme_tpl()
}
\value{

}
\description{
This function sets a custom theme for visualizations made with ggplot in line with the TPL style guide
If any trouble loading fonts, install and load the following package:
install.packages("extrafont")
library(extrafont)
}
