% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tpl_legend_top.R
\name{theme_tpl_legend_top}
\alias{theme_tpl_legend_top}
\title{Customize a ggplot theme}
\usage{
theme_tpl_legend_top()
}
\value{

}
\description{
This function sets a custom theme in line with the TPL style guide for visualizations made with ggplot. This sets a legend horizontally across the top of the plot. It can be layered on to any ggplot visualization just like a built-in theme.
}
\details{
If any trouble loading fonts, install and load the following package:
install.packages("extrafont")
library(extrafont)
}
