% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_colors.R
\name{tpl_colors}
\alias{tpl_colors}
\title{Load TPL brand colors}
\usage{
tpl_colors(named = TRUE)
}
\arguments{
\item{named}{Boolean that indicates whether vector should be named}
}
\value{
A vector of hex codes, which can be accessed by the name of the color if named=T
}
\description{
Function that loads TPL's brand colors as a vector of hex codes. Can be used to set colors in plots and charts.
}
\examples{
colors <- tpl_colors()
colors['darkblue'] #to access the named object
colors[['darkblue']] #to access the hex code directly
}
